+++ Sources/Swift/Integrations/WatchdogTerminations/SentryWatchdogTerminationTrackingIntegration.swift
@@ -3,13 +3,17 @@ import Foundation
 
 #if (os(iOS) || os(tvOS) || os(visionOS)) && !SENTRY_NO_UI_FRAMEWORK
 
-typealias WatchdogTerminationTrackingProvider = ANRTrackerBuilder & ProcessInfoProvider & AppStateManagerProvider & WatchdogTerminationScopeObserverBuilder & WatchdogTerminationTrackerBuilder
+typealias WatchdogTerminationTrackingProvider = ProcessInfoProvider & HangTrackerProvider & AppStateManagerProvider & WatchdogTerminationScopeObserverBuilder & WatchdogTerminationTrackerBuilder
 
-final class SentryWatchdogTerminationTrackingIntegration<Dependencies: WatchdogTerminationTrackingProvider>: NSObject, SwiftIntegration, SentryANRTrackerDelegate {
+final class SentryWatchdogTerminationTrackingIntegration<Dependencies: WatchdogTerminationTrackingProvider>: NSObject, SwiftIntegration {
 
     private let tracker: SentryWatchdogTerminationTracker
-    private let anrTracker: SentryANRTracker
+    private let timeoutInterval: TimeInterval
+    private let hangTracker: HangTracker
     private let appStateManager: SentryAppStateManager
+    
+    private var hasStartedHang: Bool = false
+    private var callbackId: UUID?
 
     init?(with options: Options, dependencies: Dependencies) {
         guard options.enableWatchdogTerminationTracking else {
@@ -33,13 +37,24 @@ final class SentryWatchdogTerminationTrackingIntegration<Dependencies: WatchdogT
         }
 
         tracker = terminationTracker
-        anrTracker = dependencies.getANRTracker(options.appHangTimeoutInterval)
+        timeoutInterval = options.appHangTimeoutInterval
+        hangTracker = dependencies.hangTracker
         appStateManager = dependencies.appStateManager
 
         super.init()
 
         tracker.start()
-        anrTracker.add(listener: self)
+        callbackId = hangTracker.addOngoingHangObserver { [weak self] interval, ongoing in
+            guard let self, interval > timeoutInterval else {
+                return
+            }
+
+            if ongoing {
+                hangStarted()
+            } else {
+                hangStopped()
+            }
+        }
 
         let scopeObserver = dependencies.getWatchdogTerminationScopeObserverWithOptions(options)
 
@@ -68,18 +83,24 @@ final class SentryWatchdogTerminationTrackingIntegration<Dependencies: WatchdogT
 
     func uninstall() {
         tracker.stop()
-        anrTracker.remove(listener: self)
+        
+        guard let callbackId else {
+            return
+        }
+        hangTracker.removeObserver(id: callbackId)
     }
 
-    // MARK: - SentryANRTrackerDelegate
+    func hangStarted() {
+        guard !hasStartedHang else { return }
 
-    func anrDetected(type: SentryANRType) {
+        hasStartedHang = true
         appStateManager.updateAppState { appState in
             appState.isANROngoing = true
         }
     }
 
-    func anrStopped(result: SentryANRStoppedResult?) {
+    func hangStopped() {
+        hasStartedHang = false
         appStateManager.updateAppState { appState in
             appState.isANROngoing = false
         }